Struct atoc2gtfs::setup::Config

pub struct Config<'a> {
    pub input_path: &'a Path,
    pub output_path: &'a Path,
}

Capture and collect the runtime configuration

Fields
input_path: &'a Path

path to the input file

output_path: &'a Path

path to write output file

Implementations

impl<'a> Config<'a>

pub fn build_from_cli( parsed_args: &[String] ) -> Result<Config<'_>, Box<dyn Error>>

Builds an instance of Config after parsing CLI inputs. Used to collect input and output paths, check the input exists and is a file, and that both both the input and outpus are zip files. Also checks the required ATOC files are within the input zip. Both .zip and .ZIP are valid extensions.

Arguments

parsed_args - A vector of strings denoting the parsed inputs. Expecting the format: [BINARY_NAME, ATOC_PATH, OUTPUT_PATH] (this is the result of calling std::env::args().collect())

Examples

use atoc2gtfs::setup::Config;
let dummy_parsed_args = vec![
    "".to_string(),                 // empty dummy binary name (not used)
    "./tests/data/dummy_atoc.zip".to_string(),  // dummy zipped atoc file
    "dummy_output.zip".to_string(),    // dummy output file path
];
let config = Config::build_from_cli(&dummy_parsed_args);
assert!(config.is_ok());

See Also

Config::new - Create a Config instance by supplying arguments directly.

pub fn new( input_path: &'a str, output_path: &'a str ) -> Result<Config<'a>, Box<dyn Error>>

Create an instance of Config. Used to collect input and output paths, check the input exists and is a file, and that both both the input and outpus are zip files. Also checks the required ATOC files are within the input zip. Both .zip and .ZIP are valid extensions.

Arguments

input_path - Path to the input ATOC zip file to be converted, as a str.

output_path - Path to user to save the converted GTFS file, as a str.

Examples

use atoc2gtfs::setup::Config;
let input_path = "./tests/data/dummy_atoc.zip";
let output_path = "dummy_output.zip";
let config = Config::new(input_path, output_path);
assert!(config.is_ok());

See Also

Config::build_from_cli - Building Config by parsing CLI arguments.