Function atoc2gtfs::utils::io::check_extension

pub fn check_extension(
    path: &Path,
    extensions: &Vec<&str>
) -> Result<(), String>

Checks if the provded input path has the expected file extension.

Returns a Result enmum as follows:

OK(()) when the provided path has the expected file extension.

Er() in all other cases. This could be due to the path having no file extension (say a directory) or it is not a zip file (say .txt). The Er() will contain a String error message with more details.

Note: this function wraps std::path::Path::extension. See this method’s documentation for more details.

Arguments

path - Path to input file to test.

extensions - A vector of &str types containing valid extensions.

Example (checking for a ZIP archive file)

use atoc2gtfs::utils::io;
use std::path::Path;

let zip_example = Path::new("example.zip");             // is a zip file
let text_example = Path::new("example.txt");            // is not a zip file
let accept_zip_exts: Vec<&str> = vec!["zip", "ZIP"];    // valid extensions
assert!(io::check_extension(&zip_example, &accept_zip_exts).is_ok());
assert!(io::check_extension(&text_example, &accept_zip_exts).is_err());